Refactored suggestions
