+++ web/src/components/ThemeContext.tsx
@@ -4,7 +4,6 @@ import { ThemeProvider } from 'styled-components'
 import { UiDirectionType } from 'translations'
 
 import buildConfig from '../constants/buildConfig'
-import { contrastThemeMediaQueries } from '../constants/contrastThemeMediaQueries'
 
 type ThemeType = 'light' | 'contrast'
 
@@ -22,6 +21,12 @@ export const ThemeContext = createContext<ThemeContextType>({
   toggleTheme: () => undefined,
 })
 
+export const contrastThemeMediaQueries = [
+  '(forced-colors: active)' /* to detect enabled forced colors mode: https://developer.mozilla.org/en-US/docs/Web/CSS/@media/forced-colors */,
+  '(prefers-contrast: more)' /* to detect a lower/higher contrast: https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-contrast */,
+  '(prefers-color-scheme: dark)' /* is used to detect if a user has requested light or dark color themes: https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-color-scheme */,
+].map(query => window.matchMedia(query))
+
 const getSystemTheme = (): ThemeType => (contrastThemeMediaQueries.some(query => query.matches) ? 'contrast' : 'light')
 
 type ThemeContainerProps = {
@@ -30,20 +35,17 @@ type ThemeContainerProps = {
 }
 
 export const ThemeContainer = ({ children, contentDirection }: ThemeContainerProps): ReactElement => {
-  const [themeType, setThemeType] = useState<ThemeType>(() => getSystemTheme())
+  const [themeType, setThemeType] = useState<ThemeType>(getSystemTheme)
 
   useEffect(() => {
-    const handleSystemChange = () => {
-      const systemTheme = getSystemTheme()
-      setThemeType(prev => (prev !== systemTheme ? systemTheme : prev))
-    }
+    const updateTheme = () => setThemeType(getSystemTheme())
 
-    handleSystemChange()
+    updateTheme()
 
-    contrastThemeMediaQueries.forEach(query => query.addEventListener('change', handleSystemChange))
+    contrastThemeMediaQueries.forEach(query => query.addEventListener('change', updateTheme))
 
     return () => {
-      contrastThemeMediaQueries.forEach(query => query.removeEventListener('change', handleSystemChange))
+      contrastThemeMediaQueries.forEach(query => query.removeEventListener('change', updateTheme))
     }
   }, [])
